requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/City-of-Helsinki/django-helusers.git@v0.1#egg=django-helusers
-e git+https://github.com/City-of-Helsinki/wagtail-svgmap.git@v0.1.1#egg=wagtail_svgmap
-e git+https://github.com/City-of-Helsinki/wagtailtinymce.git@extensibility#egg=wagtailtinymce==4.4.3.0
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
autopep8==1.2.4
beautifulsoup4==4.5.1     # via wagtail
billiard==3.3.0.23        # via celery
celery==3.1.23
click==6.6                # via pip-tools
contextlib2==0.5.4        # via raven
defusedxml==0.4.1         # via python3-openid
django-allauth==0.27.0
django-appconf==1.0.2     # via django-compressor
django-bower==5.2.0
django-celery==3.1.17
django-compressor-autoprefixer==0.1.0
django-compressor==2.1
django-libsass==0.7
django-modelcluster==2.0  # via wagtail
django-taggit==0.18.3     # via wagtail
django-treebeard==4.0.1   # via wagtail
Django==1.9.10            # via django-allauth, django-bower, django-treebeard, wagtail
djangorestframework==3.4.7  # via wagtail
elasticsearch==2.4.0
first==2.0.1              # via pip-tools
html5lib==0.999999        # via wagtail
isort==4.2.5
Jinja2==2.8
kombu==3.0.35             # via celery
libsass==0.11.1           # via django-libsass
MarkupSafe==0.23          # via jinja2
oauthlib==2.0.0           # via requests-oauthlib
pep8==1.7.0               # via autopep8
Pillow==3.3.1             # via wagtail
pip-tools==1.7.0
psycopg2==2.6.2
py==1.4.31                # via pytest
pytest-django==3.0.0
pytest==3.0.2
python3-openid==3.0.10    # via django-allauth
pytz==2016.6.1            # via celery, django-modelcluster
raven==5.27.1
rcssmin==1.0.6            # via django-compressor
redis==2.10.5
requests-oauthlib==0.7.0  # via django-allauth, tweepy
requests==2.11.1          # via django-allauth, requests-oauthlib, tweepy
rjsmin==1.0.12            # via django-compressor
six==1.10.0               # via django-bower, html5lib, libsass, pip-tools, tweepy
tweepy==3.5.0
Unidecode==0.4.19         # via wagtail
urllib3==1.18             # via elasticsearch
wagtail-blog==1.6.7
wagtail==1.6.2
Willow==0.3.1             # via wagtail
